# scripts/docs/openapi-full.json
{"openapi":"3.1.0","info":{"title":"SwiftFrame API 接口文档","description":"基于 Spring Boot 3 + SpringDoc 的后端架构接口文档<br/><br/>### 响应状态码说明\n| 状态码 | 描述 |\n| :--- | :--- |\n| PUB-200-00 | 操作成功 |\n| PUB-500-00 | 服务器内部错误 |\n| PUB-503-00 | 服务不可用 |\n| PUB-500-01 | 数据库操作异常 |\n| PUB-500-02 | 加解密异常 |\n| PUB-505-00 | 网络通信异常 |\n| PUB-404-00 | 资源不存在 |\n| BUS-400-00 | 请求参数错误 |\n| BUS-401-00 | 未授权访问 |\n| BUS-403-00 | 禁止访问 |\n| BUS-404-00 | 资源不存在 |\n| BUS-405-00 | 请求方法不允许 |\n| BUS-405-00 | 业务不允许 |\n","contact":{"name":"star","email":"star@example.com"},"license":{"name":"Apache 2.0","url":"https://springdoc.org"},"version":"v1.0"},"servers":[{"url":"http://localhost:8081/swift","description":"Generated server url"}],"security":[{"JWT":[]}],"tags":[{"name":"角色管理","description":"系统角色相关接口"},{"name":"鉴权管理","description":"登录和注册接口"},{"name":"用户管理","description":"系统用户相关接口"},{"name":"系统监控","description":"系统监控数据接口"}],"paths":{"/user/{userId}":{"get":{"tags":["用户管理"],"summary":"获取用户详情","description":"根据ID获取用户的详细信息","operationId":"getUserById","parameters":[{"name":"userId","in":"path","description":"用户ID","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PubResultObject"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PubResultSwiftUserDetails"}}}}}},"put":{"tags":["用户管理"],"summary":"更新用户","description":"更新指定用户的信息","operationId":"updateUser","parameters":[{"name":"userId","in":"path","description":"用户ID","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SwiftUserDetails"}}},"required":true},"responses":{"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PubResultObject"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PubResultSwiftUserDetails"}}}}}},"delete":{"tags":["用户管理"],"summary":"删除用户","description":"删除指定的用户","operationId":"deleteUser","parameters":[{"name":"userId","in":"path","description":"用户ID","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PubResultObject"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PubResultVoid"}}}}}}},"/user/{userId}/unlock":{"put":{"tags":["用户管理"],"summary":"解锁用户","description":"解锁指定的用户账户","operationId":"unlockUser","parameters":[{"name":"userId","in":"path","description":"用户ID","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PubResultObject"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PubResultVoid"}}}}}}},"/user/{userId}/password":{"put":{"tags":["用户管理"],"summary":"修改密码","description":"修改指定用户的密码","operationId":"changePassword","parameters":[{"name":"userId","in":"path","description":"用户ID","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"type":"string","description":"新密码"}}},"required":true},"responses":{"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PubResultObject"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PubResultSwiftUserDetails"}}}}}}},"/user/{userId}/lock":{"put":{"tags":["用户管理"],"summary":"锁定用户","description":"锁定指定的用户账户","operationId":"lockUser","parameters":[{"name":"userId","in":"path","description":"用户ID","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PubResultObject"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PubResultVoid"}}}}}}},"/user/{userId}/enable":{"put":{"tags":["用户管理"],"summary":"启用用户","description":"启用指定的用户账户","operationId":"enableUser","parameters":[{"name":"userId","in":"path","description":"用户ID","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PubResultObject"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PubResultVoid"}}}}}}},"/user/{userId}/disable":{"put":{"tags":["用户管理"],"summary":"禁用用户","description":"禁用指定的用户账户","operationId":"disableUser","parameters":[{"name":"userId","in":"path","description":"用户ID","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PubResultObject"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PubResultVoid"}}}}}}},"/roles/{roleId}":{"get":{"tags":["角色管理"],"summary":"获取角色详情","description":"根据ID获取角色的详细信息","operationId":"getRoleById","parameters":[{"name":"roleId","in":"path","description":"角色ID","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PubResultObject"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PubResultSwiftRole"}}}}}},"put":{"tags":["角色管理"],"summary":"更新角色","description":"更新指定角色的信息","operationId":"updateRole","parameters":[{"name":"roleId","in":"path","description":"角色ID","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SwiftRole"}}},"required":true},"responses":{"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PubResultObject"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PubResultSwiftRole"}}}}}},"delete":{"tags":["角色管理"],"summary":"删除角色","description":"删除指定的角色（不能删除已分配用户的角色）","operationId":"deleteRole","parameters":[{"name":"roleId","in":"path","description":"角色ID","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PubResultObject"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PubResultVoid"}}}}}}},"/user":{"post":{"tags":["用户管理"],"summary":"创建用户","description":"创建新的系统用户","operationId":"createUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SwiftUserDetails"}}},"required":true},"responses":{"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PubResultObject"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PubResultSwiftUserDetails"}}}}}}},"/user/{userId}/roles/{roleId}":{"post":{"tags":["用户管理"],"summary":"分配角色","description":"将指定角色分配给用户","operationId":"assignRoleToUser","parameters":[{"name":"userId","in":"path","description":"用户ID","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"roleId","in":"path","description":"角色ID","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PubResultObject"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PubResultVoid"}}}}}},"delete":{"tags":["用户管理"],"summary":"收回角色","description":"从用户中收回指定的角色","operationId":"removeRoleFromUser","parameters":[{"name":"userId","in":"path","description":"用户ID","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"roleId","in":"path","description":"角色ID","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PubResultObject"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PubResultVoid"}}}}}}},"/roles":{"get":{"tags":["角色管理"],"summary":"获取所有角色","description":"获取系统中所有的角色列表","operationId":"getAllRoles","responses":{"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PubResultObject"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PubResultListSwiftRole"}}}}}},"post":{"tags":["角色管理"],"summary":"创建角色","description":"创建新的系统角色","operationId":"createRole","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SwiftRole"}}},"required":true},"responses":{"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PubResultObject"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PubResultSwiftRole"}}}}}}},"/roles/{roleId}/authorities/{authorityId}":{"post":{"tags":["角色管理"],"summary":"授予权限","description":"将指定权限授予给角色","operationId":"grantAuthorityToRole","parameters":[{"name":"roleId","in":"path","description":"角色ID","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"authorityId","in":"path","description":"权限ID","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PubResultObject"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PubResultVoid"}}}}}},"delete":{"tags":["角色管理"],"summary":"收回权限","description":"从角色中收回指定的权限","operationId":"revokeAuthorityFromRole","parameters":[{"name":"roleId","in":"path","description":"角色ID","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"authorityId","in":"path","description":"权限ID","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PubResultObject"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PubResultVoid"}}}}}}},"/auth/register":{"post":{"tags":["鉴权管理"],"summary":"用户注册","description":"注册新用户","operationId":"registerUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterRequest"}}},"required":true},"responses":{"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PubResultObject"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PubResultSwiftUserDetails"}}}}}}},"/auth/logout":{"post":{"tags":["鉴权管理"],"summary":"用户登出","description":"用户退出登录，使token失效","operationId":"logout","parameters":[{"name":"Authorization","in":"header","required":true,"schema":{"type":"string"}}],"responses":{"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PubResultObject"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PubResultString"}}}}}}},"/auth/login":{"post":{"tags":["鉴权管理"],"summary":"用户登录","description":"通过用户名和密码获取 JWT Token","operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PubResultObject"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PubResultJwtToken"}}}}}}},"/auth/forgot-password/send-code":{"post":{"tags":["鉴权管理"],"summary":"发送忘记密码验证码","description":"向邮箱发送密码重置验证码","operationId":"sendForgotPasswordCode","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ForgotPasswordRequest"}}},"required":true},"responses":{"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PubResultObject"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PubResultString"}}}}}}},"/auth/forgot-password/reset":{"post":{"tags":["鉴权管理"],"summary":"重置密码","description":"使用验证码重置密码","operationId":"resetPassword","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResetPasswordRequest"}}},"required":true},"responses":{"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PubResultObject"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PubResultString"}}}}}}},"/user/{userId}/roles":{"get":{"tags":["用户管理"],"summary":"获取用户角色","description":"获取指定用户拥有的所有角色","operationId":"getUserRoles","parameters":[{"name":"userId","in":"path","description":"用户ID","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PubResultObject"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PubResultSetSwiftRole"}}}}}}},"/user/info":{"get":{"tags":["用户管理"],"summary":"获取当前用户信息","description":"获取当前登录用户的详细信息","operationId":"getUserInfo","responses":{"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PubResultObject"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PubResultMapStringObject"}}}}}}},"/user/all":{"get":{"tags":["用户管理"],"summary":"获取所有用户","description":"获取系统中所有的用户列表","operationId":"getAllUsers","responses":{"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PubResultObject"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PubResultListSwiftUserDetails"}}}}}}},"/roles/{roleId}/users":{"get":{"tags":["角色管理"],"summary":"获取角色用户","description":"获取拥有指定角色的所有用户","operationId":"getUsersWithRole","parameters":[{"name":"roleId","in":"path","description":"角色ID","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PubResultObject"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PubResultSetSwiftUserDetails"}}}}}}},"/roles/{roleId}/authorities":{"get":{"tags":["角色管理"],"summary":"获取角色权限","description":"获取指定角色拥有的所有权限","operationId":"getRoleAuthorities","parameters":[{"name":"roleId","in":"path","description":"角色ID","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PubResultObject"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PubResultSetSwiftAuthority"}}}}}}},"/monitor/thread":{"get":{"tags":["系统监控"],"summary":"线程信息","description":"获取线程相关信息","operationId":"getThreadDetails","responses":{"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PubResultObject"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PubResultMapStringObject"}}}}}}},"/monitor/runtime":{"get":{"tags":["系统监控"],"summary":"运行时信息","description":"获取应用运行时信息","operationId":"getRuntimeDetails","responses":{"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PubResultObject"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PubResultMapStringObject"}}}}}}},"/monitor/overview":{"get":{"tags":["系统监控"],"summary":"系统概览","description":"获取系统整体概览信息","operationId":"getSystemOverview","responses":{"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PubResultObject"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PubResultMapStringObject"}}}}}}},"/monitor/os":{"get":{"tags":["系统监控"],"summary":"操作系统信息","description":"获取操作系统详细信息","operationId":"getOsDetails","responses":{"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PubResultObject"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PubResultMapStringObject"}}}}}}},"/monitor/memory":{"get":{"tags":["系统监控"],"summary":"内存详情","description":"获取系统内存使用详情","operationId":"getMemoryDetails","responses":{"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PubResultObject"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PubResultMapStringObject"}}}}}}},"/monitor/jvm":{"get":{"tags":["系统监控"],"summary":"JVM详情","description":"获取JVM运行时详细信息","operationId":"getJvmDetails","responses":{"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PubResultObject"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PubResultMapStringObject"}}}}}}},"/monitor/all":{"get":{"tags":["系统监控"],"summary":"全部监控数据","description":"获取所有监控数据的聚合接口","operationId":"getAllMonitorData","responses":{"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PubResultObject"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PubResultMapStringObject"}}}}}}},"/auth/publicKey":{"get":{"tags":["鉴权管理"],"summary":"获取RSA公钥","description":"获取用于前端加密密码的RSA公钥","operationId":"getPublicKey","responses":{"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PubResultObject"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PubResultString"}}}}}}}},"components":{"schemas":{"PubResultObject":{"type":"object","properties":{"code":{"type":"string"},"msg":{"type":"string"},"data":{}}},"GrantedAuthority":{"type":"object","properties":{"authority":{"type":"string"}}},"SwiftAuthority":{"type":"object","properties":{"authorityId":{"type":"integer","format":"int64"},"name":{"type":"string"},"description":{"type":"string"}}},"SwiftRole":{"type":"object","properties":{"roleId":{"type":"integer","format":"int64"},"name":{"type":"string"},"description":{"type":"string"},"roleAuthorities":{"type":"array","items":{"$ref":"#/components/schemas/SwiftRoleAuthority"},"uniqueItems":true}}},"SwiftRoleAuthority":{"type":"object","properties":{"id":{"$ref":"#/components/schemas/SwiftRoleAuthorityId"},"role":{"$ref":"#/components/schemas/SwiftRole"},"authority":{"$ref":"#/components/schemas/SwiftAuthority"},"grantedAt":{"type":"string","format":"date-time"}}},"SwiftRoleAuthorityId":{"type":"object","properties":{"roleId":{"type":"integer","format":"int64"},"authorityId":{"type":"integer","format":"int64"}}},"SwiftUserDetails":{"type":"object","properties":{"userId":{"type":"integer","format":"int64"},"username":{"type":"string","maxLength":20,"minLength":4,"pattern":"^[a-zA-Z0-9_]+$"},"fullName":{"type":"string"},"email":{"type":"string"},"phone":{"type":"string","pattern":"^1[3-9]\\d{9}$"},"enabled":{"type":"boolean"},"accountNonExpired":{"type":"boolean"},"accountNonLocked":{"type":"boolean"},"credentialsNonExpired":{"type":"boolean"},"failedLoginAttempts":{"type":"integer","format":"int32"},"lockUntil":{"type":"string","format":"date-time"},"passwordChangedAt":{"type":"string","format":"date-time"},"lastLoginAt":{"type":"string","format":"date-time"},"lastLoginIp":{"type":"string"},"createdAt":{"type":"string","format":"date-time"},"userRoles":{"type":"array","items":{"$ref":"#/components/schemas/SwiftUserRole"},"uniqueItems":true},"authorities":{"type":"array","items":{"$ref":"#/components/schemas/GrantedAuthority"}}},"required":["email","username"]},"SwiftUserRole":{"type":"object","properties":{"userId":{"$ref":"#/components/schemas/SwiftUserRoleId"},"user":{"$ref":"#/components/schemas/SwiftUserDetails"},"role":{"$ref":"#/components/schemas/SwiftRole"},"assignedAt":{"type":"string","format":"date-time"},"assignedBy":{"type":"string"}}},"SwiftUserRoleId":{"type":"object","properties":{"userId":{"type":"integer","format":"int64"},"roleId":{"type":"integer","format":"int64"}}},"PubResultSwiftUserDetails":{"type":"object","properties":{"code":{"type":"string"},"msg":{"type":"string"},"data":{"$ref":"#/components/schemas/SwiftUserDetails"}}},"PubResultVoid":{"type":"object","properties":{"code":{"type":"string"},"msg":{"type":"string"},"data":{"type":"object"}}},"PubResultSwiftRole":{"type":"object","properties":{"code":{"type":"string"},"msg":{"type":"string"},"data":{"$ref":"#/components/schemas/SwiftRole"}}},"RegisterRequest":{"type":"object","properties":{"username":{"type":"string","maxLength":20,"minLength":4,"pattern":"^[a-zA-Z0-9_]+$"},"password":{"type":"string","maxLength":30,"minLength":8,"pattern":"^(?=.*[a-z])(?=.*[A-Z])(?=.*\\d)(?=.*[@$!%*?&#_+])[A-Za-z\\d@$!%*?&#_+]+$"},"email":{"type":"string"},"phone":{"type":"string","pattern":"^1[3-9]\\d{9}$"}},"required":["email","password","username"]},"PubResultString":{"type":"object","properties":{"code":{"type":"string"},"msg":{"type":"string"},"data":{"type":"string"}}},"LoginRequest":{"type":"object","properties":{"username":{"type":"string","maxLength":20,"minLength":4,"pattern":"^[a-zA-Z0-9_]+$"},"password":{"type":"string","maxLength":30,"minLength":8,"pattern":"^(?=.*[a-z])(?=.*[A-Z])(?=.*\\d)(?=.*[@$!%*?&])[A-Za-z\\d@$!%*?&]+$"}},"required":["password","username"]},"JwtToken":{"type":"object","properties":{"accessToken":{"type":"string"},"refreshToken":{"type":"string"}}},"PubResultJwtToken":{"type":"object","properties":{"code":{"type":"string"},"msg":{"type":"string"},"data":{"$ref":"#/components/schemas/JwtToken"}}},"ForgotPasswordRequest":{"type":"object","properties":{"email":{"type":"string"}},"required":["email"]},"ResetPasswordRequest":{"type":"object","properties":{"email":{"type":"string"},"verificationCode":{"type":"string"},"newPassword":{"type":"string","maxLength":30,"minLength":8,"pattern":"^(?=.*[a-z])(?=.*[A-Z])(?=.*\\d)(?=.*[@$!%*?&#_+])[A-Za-z\\d@$!%*?&#_+]+$"}},"required":["email","newPassword","verificationCode"]},"PubResultSetSwiftRole":{"type":"object","properties":{"code":{"type":"string"},"msg":{"type":"string"},"data":{"type":"array","items":{"$ref":"#/components/schemas/SwiftRole"},"uniqueItems":true}}},"PubResultMapStringObject":{"type":"object","properties":{"code":{"type":"string"},"msg":{"type":"string"},"data":{"type":"object","additionalProperties":{}}}},"PubResultListSwiftUserDetails":{"type":"object","properties":{"code":{"type":"string"},"msg":{"type":"string"},"data":{"type":"array","items":{"$ref":"#/components/schemas/SwiftUserDetails"}}}},"PubResultListSwiftRole":{"type":"object","properties":{"code":{"type":"string"},"msg":{"type":"string"},"data":{"type":"array","items":{"$ref":"#/components/schemas/SwiftRole"}}}},"PubResultSetSwiftUserDetails":{"type":"object","properties":{"code":{"type":"string"},"msg":{"type":"string"},"data":{"type":"array","items":{"$ref":"#/components/schemas/SwiftUserDetails"},"uniqueItems":true}}},"PubResultSetSwiftAuthority":{"type":"object","properties":{"code":{"type":"string"},"msg":{"type":"string"},"data":{"type":"array","items":{"$ref":"#/components/schemas/SwiftAuthority"},"uniqueItems":true}}}},"securitySchemes":{"JWT":{"type":"http","name":"Authorization","in":"header","scheme":"bearer","bearerFormat":"JWT"}}}}